+++ km3pipe/tests/test_mc.py
@@ -23,7 +23,7 @@ class TestMc(TestCase):
 
     def test_pdg2name(self):
         self.assertEqual('mu-', pdg2name(13))
-        self.assertEqual('anu_tau', pdg2name(-16))
+        self.assertEqual('nu(tau)~', pdg2name(-16))
 
     def test_pdg2name_returns_NA_for_unknown_particle(self):
         self.assertEqual('N/A', pdg2name(0))